requirements.txt
pytest==2.9.0
pytest-xdist==1.17.1
pytest-allure-adaptor==1.7.10
flaky==3.3.0
selenium>=3.1.0
requests==2.18.4